Move jax.nn implementation into jax._src.nn.
