+++ jax/_src/nn/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2019 Google LLC
+# Copyright 2020 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,29 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-"""
-Common neural network layer initializers, consistent with definitions
-used in Keras and Sonnet.
-"""
-
-# flake8: noqa: F401
-from jax._src.nn.initializers import (
-  delta_orthogonal,
-  glorot_normal,
-  glorot_uniform,
-  he_normal,
-  he_uniform,
-  kaiming_normal,
-  kaiming_uniform,
-  lecun_normal,
-  lecun_uniform,
-  normal,
-  ones,
-  orthogonal,
-  uniform,
-  variance_scaling,
-  xavier_normal,
-  xavier_uniform,
-  zeros,
-)
